migrate-profiler-mixin-move: Move profiler controls to SchedulerProfilerManager
Mechanical cut + paste for the ``migrate-profiler-mixin`` mech move.

Cut the @staticmethods (``_init_profile`` / ``_start_profile`` /
``_merge_profile_traces`` / ``_stop_profile`` / ``_profile_batch_predicate``
/ ``_profile``) from ``scheduler_profiler_mixin.py`` and paste them into
``SchedulerProfilerManager`` class body in
``scheduler_components/profiler_manager.py``. The NPU platform-patch
block + module logger relocate alongside as supporting module-level
statements. The source file is deleted; ``SchedulerProfilerMixin`` is
dropped from the Scheduler inheritance list and its import from
``scheduler.py`` is removed.

Method bodies otherwise byte-identical. ``@staticmethod`` decorators
dropped; ``self: "SchedulerProfilerManager"`` annotation simplified to
bare ``self``.

Callers updated (pure prefix transformation):
  ``self._profile_batch_predicate(self.profiler_manager, ...)`` →
  ``self.profiler_manager._profile_batch_predicate(...)``
  ``self._profile(self.profiler_manager, ...)`` →
  ``self.profiler_manager._profile(...)``

Test fixture rewires ``SchedulerProfilerMixin._init_profile`` import to
``SchedulerProfilerManager._init_profile`` at its new path.

Refactor chain ID: migrate-profiler-mixin-move
